infoschema: migrate `dataForTables` and `dataForPartitions` from package `infoschema` to `executor` #15034 #15037 (#15208)
